refact/utils function
